some tests, refactor

+++ src/main/java/pl/witomir/ceneov2/currency/api/ExchangeRateModelInterface.java
@@ -3,5 +3,5 @@ package pl.witomir.ceneov2.currency.api;
 public interface ExchangeRateModelInterface {
     public String getCode();
 
-    public Float getMid();
+    public Double getMid();
 }
