+++ src/main/java/pl/witomir/ceneov2/currency/api/nbp/model/NbpRate.java
@@ -5,7 +5,7 @@ import pl.witomir.ceneov2.currency.api.ExchangeRateModelInterface;
 public class NbpRate implements ExchangeRateModelInterface {
     private String code;
     private String currency;
-    private Float mid;
+    private Double mid;
 
     public String getCode() {
         return code;
@@ -23,11 +23,11 @@ public class NbpRate implements ExchangeRateModelInterface {
         this.currency = currency;
     }
 
-    public Float getMid() {
+    public Double getMid() {
         return mid;
     }
 
-    public void setMid(Float mid) {
+    public void setMid(Double mid) {
         this.mid = mid;
     }
 }
